improvement: split igniter generators into per-strategy and per-add-on tasks
Break the monolithic `ash_authentication.add_strategy` and
`ash_authentication.add_add_on` tasks into composable sub-tasks:

- `ash_authentication.add_strategy.password`
- `ash_authentication.add_strategy.magic_link`
- `ash_authentication.add_strategy.api_key`
- `ash_authentication.add_add_on.audit_log`
- `ash_authentication.add_add_on.confirmation`

The parent tasks now validate the name and delegate via
`Igniter.compose_task`, so existing usage is unchanged.

Shared helpers (`ensure_identity`, `ensure_get_by_action`,
`add_remember_me_strategy`, `web_module_use_line`, `parent_module`,
`maybe_parse_module`) are added to `AshAuthentication.Igniter` as public
functions for use by third-party igniter tasks.

This makes it straightforward to add generators for new strategies and
add-ons without growing the parent task files.
